Small improvements.
- Add final to virtual override functions of HMAC libraries.
  The libraries are internal,
  we can change without effect main library ABI.
- Fix ID type of SPP class.
- Add remark of internal API to JSON parser unit test.

+++ pub/cfg.h
@@ -238,7 +238,7 @@ class Spp
         size_t mac_size;
     };
     std::map<uint32_t, key_t> m_keys;
-    int16_t m_own_id;
+    int8_t m_own_id;
   protected:
     /** @cond internal */
     bool set_val(char *line);
